+++ spec/internal/app/models/city.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+class City < ActiveRecord::Base
+  belongs_to :country
+  # attribs: id, name, country_id
+end