+++ spec/internal/app/models/continent.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+class Continent < ActiveRecord::Base
+  has_many :countries
+  # attribs: id, name
+end